Refactor account page
