+++ users/account/client/password/autoform.js
@@ -1,0 +1,23 @@
+AutoForm.hooks({
+  updatePassword: {
+    onSubmit: function(insertDoc, updateDoc, currentDoc) {
+      this.event.preventDefault();
+      var instance = this;
+      Accounts.changePassword(insertDoc.old, insertDoc.new, function(e) {
+        $('.btn-primary').attr('disabled', null);
+        if (e) {
+          // Alert the user of failure
+          sAlert.error(e.message);
+          instance.done(e.message);
+        } else {
+          // Alert the user of success
+          sAlert.success('Password Updated');
+          instance.done('Password Updated');
+          // Clear the form
+          AutoForm.resetForm('updatePassword');
+        }
+      });
+      return true;
+    }
+  }
+});
