+++ users/account/client/password/password.html
@@ -1,0 +1,13 @@
+<template name="updatePassword">
+  {{#autoForm schema=updatePasswordSchema id="updatePassword" validation="keyup"}}
+  <fieldset>
+    <legend>Change Password</legend>
+    {{> afQuickField name="old" type="password"}}
+    {{> afQuickField name="new" type="password"}}
+    {{> afQuickField name="confirm" type="password"}}
+    <div>
+      <button type="submit" class="btn btn-success">Change Password</button>
+    </div>
+  </fieldset>
+  {{/autoForm}}
+</template>
