+++ users/account/client/password/password.js
@@ -1,0 +1,8 @@
+import { UpdatePasswordSchema } from './schema';
+
+Template.updatePassword.helpers({
+  updatePasswordSchema () {
+    // Return a reference to the updatePasswordSchema
+    return UpdatePasswordSchema;
+  }
+});
