+++ users/account/client/username/username.js
@@ -1,0 +1,5 @@
+Template.setUserName.helpers({
+  user () {
+    return Meteor.user();
+  }
+});
